unused dependencies removed. folder structure organized
