+++ survey_app/build.gradle
@@ -144,14 +144,6 @@ dependencies {
     androidTestImplementation 'com.google.truth:truth:1.1.3'
     androidTestImplementation 'androidx.test:core:1.5.0'
     androidTestImplementation 'androidx.test.espresso:espresso-contrib:3.5.1'
-
-    testImplementation 'junit:junit:4.13.2'
-    testImplementation 'org.mockito:mockito-core:2.19.0'
-    testImplementation 'androidx.annotation:annotation:1.7.1'
-    testImplementation 'androidx.test:core:1.5.0'
-    testImplementation 'org.robolectric:robolectric:4.8.1'
-    testImplementation 'android.arch.core:core-testing:1.1.1'
-
 }
 
 task grantPermissionForODKXApp {
