+++ survey_app/src/androidTest/java/org/opendatakit/survey/MockFormListLoader.java
@@ -1,4 +1,4 @@
-package org.opendatakit.survey.utilities;
+package org.opendatakit.survey;
 
 import android.content.Context;
